feat: enhance SitesManagementPage with improved data table and filtering options
- Refactored SitesManagementPage to utilize a new SitesDataTable component for better organization and maintainability.
- Added CreateSiteDialog and DeleteSiteDialog components for site management functionalities.
- Implemented advanced filtering options for site owners and search functionality with clear filters capability.
- Updated English and Vietnamese localization files to include new strings for the site management features.
- Enhanced pagination controls and improved UI layout for better user experience.
