+++ messages/en.json
@@ -639,14 +639,24 @@
         "Saving": "Saving...",
         "Loading": "Loading...",
         "NoSites": "No sites found",
-        "Search": "Search sites...",
+        "Search": "Search",
+        "SearchPlaceholder": "Search sites...",
         "Showing": "Showing",
         "Previous": "Previous",
         "Next": "Next",
         "Page": "Page {current} of {total}",
-        "FilterByOwner": "Filter by owner",
+        "FilterByOwner": "Owner",
         "AllOwners": "All owners",
-        "Language": "Language"
+        "Language": "Language",
+        "FilterByOwnerTitle": "Filter by Owner",
+        "SearchOwners": "Search owners...",
+        "NoOwnersFound": "No owners found",
+        "ApplyFilter": "Apply Filter",
+        "ClearAllFilters": "Clear all filters",
+        "RemoveSearchFilter": "Remove search filter",
+        "RemoveOwnerFilter": "Remove owner filter",
+        "Of": "of",
+        "Results": "results"
       },
       "Pages": {
         "Title": "Pages - {siteName}",
