+++ messages/vi.json
@@ -631,7 +631,8 @@
         "Delete": "Xóa",
         "Pages": "Trang",
         "NoSites": "Không tìm thấy trang web nào",
-        "Search": "Tìm kiếm trang web...",
+        "Search": "Tìm kiếm",
+        "SearchPlaceholder": "Tìm kiếm trang web...",
         "Showing": "Hiển thị",
         "Previous": "Trước",
         "Page": "Trang {current} của {total}",
@@ -645,9 +646,18 @@
         "Branch": "Nhánh",
         "Status": "Trạng thái",
         "Owner": "Chủ sở hữu",
-        "FilterByOwner": "Lọc theo chủ sở hữu",
+        "FilterByOwner": "Chủ sở hữu",
         "AllOwners": "Tất cả chủ sở hữu",
-        "Language": "Ngôn ngữ"
+        "Language": "Ngôn ngữ",
+        "FilterByOwnerTitle": "Lọc theo chủ sở hữu",
+        "SearchOwners": "Tìm kiếm chủ sở hữu...",
+        "NoOwnersFound": "Không tìm thấy chủ sở hữu",
+        "ApplyFilter": "Áp dụng bộ lọc",
+        "ClearAllFilters": "Xóa tất cả bộ lọc",
+        "RemoveSearchFilter": "Xóa bộ lọc tìm kiếm",
+        "RemoveOwnerFilter": "Xóa bộ lọc chủ sở hữu",
+        "Of": "của",
+        "Results": "kết quả"
       },
       "Pages": {
         "Title": "Trang - {siteName}",
